requirements.txt
newsapi-python
requests
beautifulsoup4
faiss-cpu
openai
pandas
tqdm
python-dotenv
google-genai
yt-dlp
youtube-transcript-api
python-docx
pyarrow
pyspellchecker
dotenv
bs4
streamlit
gTTS
pymongo
fastapi
uvicorn
PyJWT
google-auth
email-validator
pytest
httpx
mongomock
